refactor: steps.ts をステップ単位ファイルに分割（advanced 4ファイル・api-practice 8ファイル）
- advanced/steps.ts (1012行) → steps/ 配下の4ファイル + index.ts に分割
- api-practice/steps.ts (1892行) → steps/ 配下の8ファイル + index.ts に分割
- index.ts が既存の export を維持するため import パス変更なし
